requirements.txt
Django==4.1.2
djangorestframework
Pillow